Section_03/requirements.txt
anyascii==0.3.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cycler==0.10.0
Flask==2.0.2
gunicorn==20.1.0
ibm-cloud-sdk-core==3.11.3
ibm-watson==5.3.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
psycopg2-binary==2.9.1
pyahocorasick==1.4.2
PyJWT==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
regex==2021.9.30
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
soupsieve==2.2.1
textsearch==0.0.21
threadpoolctl==3.0.0
tqdm==4.62.3
urllib3==1.26.7
websocket-client==1.1.0
Werkzeug==2.0.2
